challenges/misc/mission-impossible/source/pi/requirements.txt
requests
RPi.GPIO==0.7.1a4